# specs/001-price-comparison-mvp/contracts/http-api.openapi.yaml
openapi: 3.1.0
info:
  title: BestPrice MVP API
  version: 0.1.0
  description: >
    HTTP contract for the Egypt-first bilingual electronics price-comparison MVP.
    Shopper-facing endpoints exclude stale offers older than 12 hours. Admin
    endpoints are restricted to internal admins.
servers:
  - url: http://localhost:3001/api/v1
...
paths:
...
  /products/{productId}:
    get:
      tags: [Shopper]
      summary: Get product details and exact/similar groupings
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: Product not found
  /products/{productId}/offers:
    get:
      tags: [Shopper]
      summary: Get shopper-visible offers for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Offer list
          content:
            application/json:
              schema:
                type: object
                required: [productId, offers]
                properties:
                  productId:
                    type: string
                    format: uuid
                  offers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Offer'
...
  /admin/unmatched-products:
    get:
      tags: [Admin]
      summary: List unmatched or review-needed source listings
      security:
        - adminSession: []
      responses:
        '200':
          description: Review queue
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UnmatchedProduct'
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: bestprice_admin_session
  schemas:
...
    ProductDetail:
      type: object
      required:
        - productId
        - canonicalName
        - category
        - brand
        - exactOffers
        - similarProducts
      properties:
        productId:
          type: string
          format: uuid
        canonicalName:
          type: string
        canonicalNameArabic:
          type: string
          nullable: true
        category:
          type: string
        brand:
          type: string
        specs:
          type: object
          additionalProperties: true
        exactOffers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
        similarProducts:
          type: array
          items:
            $ref: '#/components/schemas/SimilarProduct'
    OfferSummary:
      type: object
      required:
        - offerId
        - store
        - priceEgp
        - availability
        - lastUpdatedAt
      properties:
        offerId:
          type: string
          format: uuid
        store:
          type: string
        priceEgp:
          type: number
        shippingEgp:
          type: number
          nullable: true
        landedPriceEgp:
          type: number
          nullable: true
        availability:
          type: string
          enum: [in_stock, limited, out_of_stock, unknown]
        lastUpdatedAt:
          type: string
          format: date-time
    Offer:
      allOf:
        - $ref: '#/components/schemas/OfferSummary'
        - type: object
          required:
            - matchLevel
            - matchConfidence
            - buyUrl
          properties:
            matchLevel:
              type: string
              enum: [exact, likely, similar]
            matchConfidence:
              type: number
              minimum: 0
              maximum: 1
            buyUrl:
              type: string
            rankingReasons:
              type: array
              items:
                type: string
    SimilarProduct:
      type: object
      required: [productId, canonicalName, matchConfidence]
      properties:
        productId:
          type: string
          format: uuid
        canonicalName:
          type: string
        matchConfidence:
          type: number
          minimum: 0
          maximum: 1
...
    UnmatchedProduct:
      type: object
      required:
        - rawProductId
        - sourceAdapterId
        - sourceTitle
        - fetchedAt
        - reviewStatus
      properties:
        rawProductId:
          type: string
          format: uuid
        sourceAdapterId:
          type: string
          format: uuid
        sourceTitle:
          type: string
        sourceUrl:
          type: string
        fetchedAt:
          type: string
          format: date-time
        reviewStatus:
          type: string
          enum: [pending, accepted, rejected]